requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.5
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.9
minio==7.2.7
SQLAlchemy==2.0.32
redis==5.0.7
slowapi==0.1.9
python-json-logger==2.0.7
sentry-sdk==1.45.0
prometheus-client==0.20.0
prefect==2.19.9
statsmodels==0.14.2
alembic==1.13.1
PyJWT==2.8.0
scipy==1.13.1
scikit-learn==1.5.2
filterpy==1.4.5
joblib==1.4.2


